+++ Backend/src/main/java/com/frank/msaphase3/repository/NoteRepository.java
@@ -3,5 +3,5 @@ package com.frank.msaphase3.repository;
 import com.frank.msaphase3.model.Note;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NoteRepository extends JpaRepository<Note, Long> {
+public interface NoteRepository extends JpaRepository<Note, Integer> {
 }
